fix(types): keep public inject void on v5/v6, expose boolean on v7
The prior commit widened Tracer.inject's declared return from void to
boolean across the public surface. That is a breaking TypeScript change:
under strict mode `const x: void = tracer.inject(...)` and returning
`tracer.inject(...)` from a void-typed function stop compiling. It must
not ship on a shipping major.

The runtime already returns the boolean and the internal AWS SDK
producers consume it through the unwrapped DatadogTracer (a plugin's
`this.tracer` resolves to `_tracer._tracer`, and the DSM checkpointer
holds that inner tracer directly), never through the public proxy. So
the boolean stays an internal signal regardless of what the public
`.d.ts` promises.

Freeze the contract per major:

1. index.d.v5.ts and the new index.d.v6.ts declare `inject(): void` —
   the current shipping majors keep their contract.
2. index.d.ts (master is 7.0.0-pre) declares `inject(): boolean` as the
   next-major public API, documented in MIGRATING.md under 6.0 to 7.0.

Generalize the release type swap: swap-v5-types.js only handled v5, so a
v6 release would publish the v7 index.d.ts. Rename it to
swap-legacy-types.js and pick index.d.v<major>.ts by the release major;
the newest major has no frozen file and is a no-op. Adding a future
frozen surface is now a single file drop. The .v<major>.ts files stay
out of the npm `files` array by design — the swap copies the chosen one
over index.d.ts at pack time, so only index.d.ts ships.

Refs: https://github.com/DataDog/dd-trace-js/pull/9183#discussion_r3540006717
