+++ package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "env": "bash ./plugin-env",
     "prepare": "node scripts/patch-istanbul-lib-coverage.js && node scripts/patch-v8-to-istanbul.js && cd vendor && npm ci --include=dev",
-    "prepack": "node scripts/release/swap-v5-types.js",
+    "prepack": "node scripts/release/swap-legacy-types.js",
     "bench": "node benchmark/index.js",
     "bench:e2e:test-optimization": "node benchmark/e2e-test-optimization/benchmark-run.js",
     "dependencies:dedupe": "yarn-deduplicate yarn.lock",
